+++ tabletest-reporter-core/src/test/java/org/tabletest/reporter/FormatListerTest.java
@@ -19,10 +19,10 @@ class FormatListerTest {
 
     @TableTest("""
         Scenario                      | Template Files                                                             | Expected Output
-        Empty template directory      | []                                                                         | [asciidoc, markdown]
-        Custom HTML and XML formats   | [table.html.peb, index.html.peb, table.xml.peb, index.xml.peb]             | [asciidoc, html, markdown, xml]
-        Formats sorted alphabetically | [table.zebra.peb, index.zebra.peb, table.aardvark.peb, index.aardvark.peb] | [aardvark, asciidoc, markdown, zebra]
-        Single custom format          | [table.custom.peb, index.custom.peb]                                       | [asciidoc, custom, markdown]
+        Empty template directory      | []                                                                         | [asciidoc, html, markdown]
+        Custom XML format             | [table.xml.peb, index.xml.peb]                                             | [asciidoc, html, markdown, xml]
+        Formats sorted alphabetically | [table.zebra.peb, index.zebra.peb, table.aardvark.peb, index.aardvark.peb] | [aardvark, asciidoc, html, markdown, zebra]
+        Single custom format          | [table.custom.peb, index.custom.peb]                                       | [asciidoc, custom, html, markdown]
         """)
     void lists_formats(
             @Scenario String scenario, List<String> templateFiles, List<String> expectedOutput, @TempDir Path tempDir)
@@ -42,7 +42,7 @@ class FormatListerTest {
     void lists_built_in_formats_when_template_directory_is_null() {
         String result = FormatLister.listFormats(null);
 
-        assertThat(result).isEqualTo("asciidoc" + System.lineSeparator() + "markdown");
+        assertThat(result).isEqualTo(String.join(System.lineSeparator(), "asciidoc", "html", "markdown"));
     }
 
     @Test
@@ -51,6 +51,6 @@ class FormatListerTest {
 
         String result = FormatLister.listFormats(nonexistent);
 
-        assertThat(result).isEqualTo("asciidoc" + System.lineSeparator() + "markdown");
+        assertThat(result).isEqualTo(String.join(System.lineSeparator(), "asciidoc", "html", "markdown"));
     }
 }